refactor(map): split legend, view-cells, and region-grow phases (Refs #4561)
Extract optional/terrain/region legend drawers, cell-grid vs unit-marker
builders, and region-growing targeting vs queue helpers while keeping
public view/render/grow APIs unchanged. Add wave-7 structural split pins.
